requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
click==8.1.3
colorama==0.4.6
Django==4.2
djangorestframework==3.14.0
dnspython==2.3.0
email-validator==2.0.0.post2
idna==3.4
joblib==1.2.0
psycopg==3.1.8
psycopg2==2.9.6
pytz==2023.3
regex==2023.3.23
sqlparse==0.4.4
tqdm==4.65.0
typing-extensions==4.5.0
tzdata==2023.3
